streamlitProj/requirements.txt
numpy==1.21.2
pandas==1.3.2
Pillow==8.4.0
scikit_learn==1.0.1
streamlit==0.87.0
nltk<3.6.5